% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_samplesize.R
\name{plot.samplesize}
\alias{plot.samplesize}
\title{Plots by mu1}
\usage{
plot.samplesize(SS.matrix,link.type)
}
\arguments{
\item{link.type}{the type of link used in the beta regression(or Wilcoxon Rank Sum Test). You can choose one or more of the following: "logit", "probit", "cloglog", "cauchit", "log", "loglog", "wilcoxon", "all"}

\item{SS.matrix}{the matrix obtained by the function samplesize.(the formula was described as the output formula in the function samplesize)}
}
\description{
Generate the comparison plots using GLM method and Wilcoxon Rank Sum Test with different mu1.
}
\details{
plot.samplesize() returns a series of plots equal to the number of mu1 used in the procedure.
Y-axis denotes minimum sample size and X-axis denotes minimum power.
}
\examples{
SSmat <- samplesize(mu0=0.56, sd0=0.255, mu1.start = 0.60, mu1.end = 0.70, mu1.by = 0.05, 
power.start = 0.7, power.end = 0.9, power.by = 0.1, link.type = "all")
plot.samplesize(SSmat, "all")
SSmat2 <- samplesize(mu0=0.56, sd0=0.255, mu1.start = 0.60, mu1.end = 0.70, mu1.by = 0.05, 
power.start = 0.7, power.end = 0.9, power.by = 0.1, link.type = c("logit","loglog","log"))
plot.samplesize(SSmat2,link.type = c("logit","loglog","log"))

}
